Complete simple find_inode function

+++ os/src/bio.rs
@@ -1,8 +1,8 @@
 use crate::virtio::blk::disk_rw;
 use fs::BLKSZ;
 
-pub fn bread(block_no: usize) -> [u8; BLKSZ] {
+pub fn bread(block_no: u32) -> [u8; BLKSZ] {
     let buf: [u8; BLKSZ] = [0; BLKSZ];
-    disk_rw(&buf, block_no * BLKSZ, false);
+    disk_rw(&buf, block_no as usize * BLKSZ, false);
     buf
 }
